requirements.txt
--upgrade hopsworks
pandas
numpy
requests
scikit-learn
xgboost
Flask
Streamlit
python-dotenv
confluent-kafka
joblib
-e .